refactor: extract chart components for better code organization
- Create TodaysSpendingChart.vue component with doughnut chart logic
- Create MonthlySpendingChart.vue component with bar chart and peak spending
- Update Graph.vue to use new modular chart components
- Remove duplicate chart logic and styling from Graph.vue
- Maintain all existing functionality and styling
- Improve code maintainability and reusability

Resolves #21
